requirements.txt
torch
numpy
scipy
einops
hydra-core
omegaconf
Pillow
tqdm
opencv-python
iopath
nibabel